src/backend/requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.1.2
bcrypt>=4.2.0
# Core Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database (SQLite - built into Python)
sqlalchemy==2.0.31

# Environment Variables
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-cov==5.0.0
httpx==0.28.1

# Code Quality
black==24.4.2
flake8==7.1.0

# File Upload
python-multipart==0.0.20
aiofiles==24.1.0
click==8.1.8

# OCR
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
fpdf==1.7.2

# Vector Database
pgvector==0.2.5

# C++ Analysis
libclang

# Gemini / Google Cloud
google-generativeai>=0.8.3
google-auth>=2.36.0

# Resilience
tenacity>=8.2.3
# Java
javalang==0.13.0
